TVIST1-237: Refactored tests and test methods to be more readable and telling
